+++ dsmr_datalogger/scripts/dsmr_datalogger_api_client.py
@@ -6,8 +6,6 @@
 """
 import datetime
 import logging
-import select
-import socket
 import time
 import re
 
@@ -19,103 +17,45 @@ import decouple
 logger = logging.getLogger('dsmrreader')
 
 
-def read_serial_port(port, baudrate, bytesize, parity, stopbits, xonxoff, rtscts, timeout, **kwargs):  # noqa: C901
-    """
-    Opens the serial port, keeps reading until we have a full telegram and yields the result to preserve the connection.
-    """
-    serial_kwargs = dict(
-        port=port,
-        baudrate=baudrate,
-        bytesize=bytesize,
-        parity=parity,
-        stopbits=stopbits,
-        xonxoff=xonxoff,
-        rtscts=rtscts,
-        timeout=timeout
-    )
-    logger.info('[%s] Opening serial port %s using options: %s', datetime.datetime.now(), port, serial_kwargs)
-    serial_handle = serial.Serial(**serial_kwargs)
+def read_telegram(url_or_port, **serial_kwargs):  # noqa: C901
+    """ Opens a serial/network connection and reads it until we have a full telegram. Yields the result """
+    MAX_BYTES_PER_READ = 512
+    MAX_READ_TIMEOUT = 1.0
 
-    telegram_start_seen = False
-    buffer = ''
-
-    while True:
-        try:
-            # We use an infinite datalogger loop and signals to break out of it. Serial
-            # operations however do not work well with interrupts, so we'll have to check for E-INTR error.
-            data = serial_handle.readline()
-        except serial.SerialException as error:
-            if str(error) == 'read failed: [Errno 4] Interrupted system call':
-                # If we were signaled to stop, we still have to finish our loop.
-                continue
-
-            # Something else and unexpected failed.
-            raise
-
-        try:
-            # Make sure weird characters are converted properly.
-            data = str(data, 'utf-8')
-        except TypeError:
-            pass
-
-        if data.startswith('/'):
-            telegram_start_seen = True
-            buffer = ''
-
-        if telegram_start_seen:
-            buffer += data
-
-        if data.startswith('!') and telegram_start_seen:
-            yield buffer
-
-
-def read_network_socket(host, port):
-    """ Opens a network socket and keeps reading until we have a full telegram. """
-    MAX_SELECT_TIMEOUT = 0.5
-    MAX_BYTES_RECV = 1024
-    MAX_BUFFER = 10 * MAX_BYTES_RECV
-
-    logger.info('[%s] Opening network socket (%s:%d)', datetime.datetime.now(), host, port)
-    socket_handle = socket.socket(
-        socket.AF_INET,  # Hardcoded ipv4, for now
-        socket.SOCK_STREAM
+    logger.info(
+        '[%s] Opening serial connection "%s" using options: %s',
+        datetime.datetime.now(),
+        url_or_port,
+        serial_kwargs
     )
 
     try:
-        socket_handle.connect((host, port))
+        serial_handle = serial.serial_for_url(url=url_or_port, timeout=MAX_READ_TIMEOUT, **serial_kwargs)
     except Exception as error:
-        raise RuntimeError('Failed to connect to network socket: {}', error)
+        raise RuntimeError('Failed to connect: {}', error)
 
     buffer = ''
 
     while True:
-        rlist, _, _ = select.select([socket_handle], [], [], MAX_SELECT_TIMEOUT)
+        incoming_bytes = serial_handle.read(MAX_BYTES_PER_READ)
+        logger.debug('[%s] Read %d Byte(s)', datetime.datetime.now(), len(incoming_bytes))
 
-        if not rlist:
+        if not incoming_bytes:
             continue
 
-        # Usually a bad sign.
-        if len(buffer) > MAX_BUFFER:
-            logger.error(
-                '[%s] Cleared buffer after reaching max size of %d Bytes!',
-                datetime.datetime.now(),
-                MAX_BUFFER
-            )
-            buffer = ''
+        incoming_data = str(incoming_bytes, 'latin_1')
 
-        incoming_bytes = socket_handle.recv(MAX_BYTES_RECV)
-        data = str(incoming_bytes, 'latin_1')
-        buffer += data
+        # Just add data to the buffer until we detect a telegram in it.
+        buffer += incoming_data
 
-        # Just add data to the buffer until we detect a telegram in it. Should work for 99% of the cases.
-        match = re.search(r'(\/.+\![A-Z0-9]{4})', buffer, re.DOTALL)
+        # Should work for 99% of the telegrams read. The checksum bits are optional due to legacy meters omitting them.
+        match = re.search(r'(\/[^/]+\![A-Z0-9]{0,4})', buffer, re.DOTALL)
 
         if not match:
             continue
 
         telegram = match.group(1)
         buffer = ''
-
         yield telegram
 
 
@@ -152,7 +92,6 @@ def main():  # noqa: C901
     logger.info('[%s] Starting...', datetime.datetime.now())
 
     # Settings.
-    DATALOGGER_SETTINGS = {}
     DATALOGGER_TIMEOUT = decouple.config('DATALOGGER_TIMEOUT', default=20, cast=float)
     DATALOGGER_SLEEP = decouple.config('DATALOGGER_SLEEP', default=0.5, cast=float)
     DATALOGGER_INPUT_METHOD = decouple.config('DATALOGGER_INPUT_METHOD')
@@ -165,33 +104,30 @@ def main():  # noqa: C901
     if len(DATALOGGER_API_HOSTS) != len(DATALOGGER_API_KEYS):
         raise RuntimeError('The number of API_HOSTS and API_KEYS given do not match each other')
 
+    serial_kwargs = {}
+
     if DATALOGGER_INPUT_METHOD == 'serial':
-        DATALOGGER_SETTINGS = dict(
-            port=decouple.config('DATALOGGER_SERIAL_PORT'),
+        serial_kwargs.update(dict(
+            url_or_port=decouple.config('DATALOGGER_SERIAL_PORT'),
             baudrate=decouple.config('DATALOGGER_SERIAL_BAUDRATE', cast=int),
             bytesize=serial.EIGHTBITS,
             parity=serial.PARITY_NONE,
             stopbits=serial.STOPBITS_ONE,
             xonxoff=1,
             rtscts=0,
-            timeout=DATALOGGER_TIMEOUT
-        )
+        ))
     elif DATALOGGER_INPUT_METHOD == 'ipv4':
-        DATALOGGER_SETTINGS = dict(
-            host=decouple.config('DATALOGGER_NETWORK_HOST'),
-            port=decouple.config('DATALOGGER_NETWORK_PORT', cast=int),
-        )
-
-    try:
-        input_function = {
-            'serial': read_serial_port,
-            'ipv4': read_network_socket,
-        }[DATALOGGER_INPUT_METHOD]
-    except KeyError:
+        serial_kwargs.update(dict(
+            url_or_port='socket://{}:{}'.format(
+                decouple.config('DATALOGGER_NETWORK_HOST'),
+                decouple.config('DATALOGGER_NETWORK_PORT', cast=int),
+            )
+        ))
+    else:
         raise RuntimeError('Unsupported DATALOGGER_INPUT_METHOD')
 
-    for telegram in input_function(**DATALOGGER_SETTINGS):
-        logger.debug('[%s] Telegram read: %s', datetime.datetime.now(), telegram)
+    for telegram in read_telegram(**serial_kwargs):
+        logger.debug("[%s] Telegram read: \n%s", datetime.datetime.now(), telegram)
 
         for current_server_index in range(len(DATALOGGER_API_HOSTS)):
             current_api_host = DATALOGGER_API_HOSTS[current_server_index]